% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{APTCopyNumberParam-class}
\alias{APTCopyNumberParam-class}
\title{Internal representation of a call to \code{apt-copynumber-cyto}}
\description{
This class encapulate all the argumnets requested by the command-line
applitaction \code{apt-copynumber-cyto} of the suit Affymetrix Power
Tools.
}

