% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APTparam.R
\name{APTparam}
\alias{APTparam}
\title{Function to create an object used to call Affymetrix Power Tools}
\usage{
APTparam(type, ...)
}
\description{
This function creates an object of class \code{APTCopyNumberParam},
\code{APTGenoQcParam} or \code{APTProbeSetGenotypeParam}. Each one 
is used to run a specific command-tool of the Affymetrix Power Tools.
}

