% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{APTProbeSetGenotypeParam-class}
\alias{APTProbeSetGenotypeParam-class}
\title{Internal representation of a call to \code{apt-geno-qc}}
\description{
This class encapulate all the argumnets requested by the command-line
applitaction \code{apt-geno-qc} of the suit Affymetrix Power Tools.
}

