% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax2apt.R
\name{Ax2APT}
\alias{Ax2APT}
\title{Function to preprocess Affymetrix Axiom arrays}
\usage{
Ax2APT(cel.files, output.path, analysis.path, xml.dqc, xml.config.1,
  xml.config.2, report.txt = "AxiomGT1.report.txt", verbose = FALSE)
}
\arguments{
\item{cel.files}{Location where the raw .CEL files are placed.}

\item{output.path}{Location where the metrix, summary and genotype file will be stored.}

\item{analysis.path}{Path until Axiom library (containing DQC and Step 1 and Step 2 XML files).}

\item{xml.dqc}{XML definition of the QC metric that evaluates the overlap between the two homozygous peaks.}

\item{report.txt}{By default \code{'AxiomGT1.report.txt'}. Output name of the QC report generated by APT.}

\item{verbose}{By default FALSE. If TRUE the function will shows messages indicating the process.}

\item{xml.config1}{Configuration in XML for the first step of Axiom QC pipeline.}

\item{xml.config2}{Configuration in XML for the first step of Aciom QC pipeline.}
}
\description{
This function preprocess the raw .CEL files from Affymetrix Axiom arrays
running the tools \code{apt-geno-qc} and \code{apt-probeset-genotype}
included into the Affymetrix Port Tools and generate the metrix, summary 
and genotype files need for \code{Ax2Mad} and \code{Ax2SnpMatrix}
}
\examples{
\dontrun{
ax2apt( "celFiles", "outputFiles", "analysisFiles", 
   "analysisFiles/Axiom_GW_Hu_SNP.r6.apt-geno-qc.AxiomQC1.xml", 
   "Axiom_GW_Hu_SNP_LessThan96_Step1.r6.apt-probeset-genotype.AxiomGT1.xml", 
   "Axiom_GW_Hu_SNP_LessThan96_Step2.r6.apt-probeset-genotype.AxiomGT1.xml",
   temppath="tempPath", verbose=TRUE )
}
}

