% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ax_snpmatrix.R
\name{Ax2SnpMatrix}
\alias{Ax2SnpMatrix}
\title{Function to create a \code{SnpMatrix} or a file compatible with \code{PLINK}}
\usage{
Ax2SnpMatrix(calls.file, annotation.file, output.type = "snpmatrix",
  output.name = NA, verbose = FALSE)
}
\arguments{
\item{calls.file}{Location where the 'intermediate' files (metrix, summary 
and genotype) generated by \code{Ax2APT}.}

\item{annotation.file}{NetAffx Annotation database file in CSV format.}

\item{output.type}{By default \code{"snpmatrix"}. Can be set to 
\code{"plink"} to generate a \code{.tped} file.}

\item{output.name}{If \code{output.type} is set to \code{"plink"}, it must 
contain the name of the output \code{.tped} file (with no extension).}

\item{verbose}{By default FALSE. If TRUE the function will shows messages 
indicating the process.}
}
\description{
This function uses the files created with \code{Ax2APT} (metrix, summary 
and genotype) to create a list containing a map file and a SnpMatrix with 
the genotype of each probe in the array. It can also be used to create 
a \code{.tped} file, compatible with \code{PLINK} (is in charge of the user 
to create the partner \code{.tfam} file).
}
\examples{
\dontrun{
# To create a SnpMatrix Container (a list with a map and a SnpMatrix)
 smc <- Ax2SnpMatrix(
 calls.file=paste0(output.path, .Platform$file.sep, "AxiomGT1.calls.txt"),
 annotation.file=paste0(path.lib, .Platform$file.sep, "Axiom_GW_Hu_SNP.na34.annot.csv"),
 output.type="snpmatrix"
 )
}
\dontrun{
# To create the .tped file compatible with PLINK
 Ax2SnpMatrix(
 calls.file=paste0(output.path, .Platform$file.sep, "AxiomGT1.calls.txt"),
 annotation.file=paste0(path.lib, .Platform$file.sep, "Axiom_GW_Hu_SNP.na34.annot.csv"),
 output.type="plink",
 output.name="ax"
 )
}
}

