% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_python.R
\name{CheckPython}
\alias{CheckPython}
\title{Function used to check the presence of python in the current system}
\usage{
CheckPython()
}
\description{
This function runs a command-line sentence to check the rpresence of
python and pandas in the current system. The run command follows:
\code{python -c "import pandas"}.
}

