% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_system.R
\name{CheckSystem}
\alias{CheckSystem}
\title{Function used to check the current system}
\usage{
CheckSystem(bits = TRUE)
}
\description{
This function checks than the current system is runs a 64bits OS. Then
checks fot python and pandas using \code{CheckPython}.
}

