% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_directory.R
\name{CopyDirectory}
\alias{CopyDirectory}
\title{Function to recursive copy of a directory}
\usage{
CopyDirectory(from, to, overwrite = TRUE, verbose = FALSE)
}
\description{
This function copies a full directory to another location
}

