% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyto2apt.R
\name{Cyto2APT}
\alias{Cyto2APT}
\title{Function to preprocess the .CEL files from Affymetrix CytoScan 750K/HD}
\usage{
Cyto2APT(arguments, verbose = FALSE)
}
\arguments{
\item{arguments}{Object of class \code{APTCopyNumberParam} with the 
'standard' or 'custom' arguments set.}

\item{verbose}{By default FALSE. If TRUE the function will shows messages 
indicating the process.}
}
\description{
This function run as a command line tool the \code{apt-copynumber-cyto}
included into the Affymetrix Port Tools to process the raw .CEL files and 
generate the \code{.cychp.txt} files that are requested by \code{Cyto2Mad}
and \code{Cyto2SnpMatrix}.
}
\examples{
\dontrun{
aptParam <- APTparam(
 type="cytoscan", 
 level="standard",
 cel.list="C:/Users/brge/Desktop/TEA_1de2", 
 output.path="C:/Users/brge/Desktop/APT_OUT", 
 analysis.path="C:/Users/brge/Desktop/lib_cytoHD",
 cdf="CytoScanHD_Array.cdf", 
 chrX="CytoScanHD_Array.chrXprobes", 
 chrY="CytoScanHD_Array.chrYprobes", 
 qca="CytoScanHD_Array.r1.qca", 
 qcc="CytoScanHD_Array.r1.qcc", 
 snp="CytoScanHD_Array.snplist.txt", 
 annot.db="CytoScanHD_Array.na32.3.annot.db", 
 refmodel="CytoScanHD_Array.na32.3.v1.REF_MODEL"
 )
 Cyto2APT(aptParam)
}
}

