% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CytoQCView.R
\name{CytoQCView}
\alias{CytoQCView}
\title{Create an object for CytoScan visualization}
\usage{
CytoQCView(path, visualization, individual, verbose = FALSE)
}
\arguments{
\item{path}{Location of the .cychp files.}

\item{visualization}{"snp", "int" or "sc".}

\item{individual}{name of the invividual for calling and sc visualizations.}

\item{verbose}{If set to \code{TRUE} usefull information is shown.}
}
\description{
This function creates an object of class \code{CytoQCViewer} that allows
to perform a visual quality control. Three plots are avaialble: snp, 
calling and sc.
}

