% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CytoQCViewer-class}
\alias{CytoQCViewer-class}
\title{Class result of the \code{CytoQCView}.}
\description{
This class encapulate all information to draw one of the three 
visualizations ('snp', 'calling' and 'sc') of Affymetrix CytoScan.
}

