% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamAPT-run.R
\docType{methods}
\name{run,APTCopyNumberParam-method}
\alias{run,APTCopyNumberParam-method}
\title{Execute \code{apt-copynumber-cyto}.}
\usage{
\S4method{run}{APTCopyNumberParam}(object)
}
\description{
Method to exceute the command-line call created with \code{APTparam} to
the command-line tool \code{apt-copynumber-cyto}.
}

