% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_convert.R
\name{update_map}
\alias{update_map}
\title{Function to update the MAP annotation of a \code{SnpMatrix Container}.}
\usage{
update_map(smc)
}
\description{
Given a \code{SnpMatrix Container}, this function returns a new 
\code{SnpMatrix Container} with its annotation (MAP slot) updates to
GRCh37.p13.
}

